restructured all game projects so they are more uniform with the way bingo does it
