+++ data/states/blackjack/__init__.py
@@ -1,0 +1,6 @@
+from . import blackjack
+
+
+def Blackjack():
+    """Return the main state for the blackjack game"""
+    return blackjack.Blackjack()
