+++ data/states/craps/__init__.py
@@ -1,0 +1,6 @@
+from . import craps
+
+
+def Craps():
+    """Return the main state for the craps game"""
+    return craps.Craps()
